extract Criterion
