+++ smarttree/_cy_column_splitter.pyi
@@ -1,6 +1,4 @@
-import numpy as np
 import pandas as pd
-from numpy.typing import NDArray
 
 from ._dataset import Dataset
 from ._types import Criterion
@@ -49,30 +47,5 @@ class CyBaseColumnSplitter:
                 \item $N_{\text{child}_i}$ — number of samples in child node;
                 \item $\text{impurity}_{\text{child}_i}$ — child node impurity.
             \end{itemize}
-        """
-        ...
-
-    def gini_index(self, mask: NDArray[np.int8]) -> float:
-        r"""
-        Calculates Gini index in a tree node.
-
-        Gini index formula in LaTeX:
-            \text{Gini Index} = 1 - \sum^C_{i=1} p_i^2
-            where
-            C - total number of classes;
-            p_i - the probability of choosing a sample with class i.
-        """
-        ...
-
-    def entropy(self, mask: NDArray[np.int8]) -> float:
-        r"""
-        Calculates entropy in a tree node.
-
-        Entropy formula in LaTeX:
-        H = \log{\overline{N}} = \sum^N_{i=1} p_i \log{(1/p_i)} = -\sum^N_{i=1} p_i \log{p_i}
-        where
-        H - entropy;
-        \overline{N} - effective number of states;
-        p_i - probability of the i-th system state.
         """
         ...
